requirements.txt
streamlit
pandas
requests
beautifulsoup4
fpdf2